+++ components/shared_code/src/shared_data_model/sources/dependency_track.py
@@ -6,6 +6,7 @@ from shared_data_model.meta.entity import Color, Entity, EntityAttribute, Entity
 from shared_data_model.meta.source import Source
 from shared_data_model.parameters import (
     URL,
+    GitHubPersonalAccessToken,
     LandingURL,
     MultipleChoiceWithAdditionParameter,
     MultipleChoiceWithDefaultsParameter,
@@ -106,6 +107,7 @@ DEPENDENCY_TRACK = Source(
             help="Only include project versions that are marked as latest.",
             metrics=["dependencies", "security_warnings", "source_up_to_dateness"],
         ),
+        "github_pat": GitHubPersonalAccessToken(metrics=["source_version"]),
     },
     entities={
         "dependencies": Entity(